requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# LLM Providers
openai>=1.6.0
google-generativeai>=0.3.2
ollama>=0.1.6

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Video/Audio Processing
moviepy>=1.0.3
faster-whisper>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
